genesis-agent/agent_backend/video-processing-service/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-video-intelligence